triage-bot-func/requirements.txt
openai
azure-functions
pyodbc